Spring boot 与 Shiro 整合

+++ src/main/java/com/amoyiki/springtest/repository/UserRepository.java
@@ -1,6 +1,5 @@
 package com.amoyiki.springtest.repository;
 
-
 import com.amoyiki.springtest.entry.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
@@ -11,6 +10,6 @@ import java.util.Optional;
  * @author: amoyiki
  * @date: 2019/2/28
  */
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUsername(String username);
 }
